+++ ChangeLog.md
@@ -28,6 +28,8 @@
   regex matching slow on glibc due to lock contention
 * [Issue #287](https://github.com/grobian/carbon-c-relay/issues/287)
   unexpected warning about rejected socket bufsize
+* [PR #282](https://github.com/grobian/carbon-c-relay/pull/282)
+  route using should not change the metric key sent
 
 
 # 3.1 (29-04-2017)